requirements.txt
python-telegram-bot[ext]>=22.0
requests>=2.28.0
python-dotenv>=1.0.0
